changed tempo normaly working
