+++ megamek/src/megamek/client/ui/swing/gameConnectionDialogs/HostDialog.java
@@ -28,7 +28,7 @@ import javax.swing.*;
 
 import megamek.client.ui.Messages;
 import megamek.client.ui.swing.util.UIUtil;
-import megamek.server.Server;
+import megamek.server.Processor.MovementProcessor;
 
 /**
  * The host game dialog shown when hosting a new game and when loading a game
@@ -150,7 +150,7 @@ public class HostDialog extends AbstractGameConnectionDialog {
     @Override
     public boolean dataValidation(String errorTitleKey) {
         try {
-            setServerPass(Server.validatePassword(getServerPass()));
+            setServerPass(MovementProcessor.validatePassword(getServerPass()));
         } catch (Exception ex) {
             JOptionPane.showMessageDialog(getOwner(), Messages.getString("MegaMek.ServerPasswordError"),
                     Messages.getString(errorTitleKey), JOptionPane.ERROR_MESSAGE);
