+++ megamek/src/megamek/common/commandline/ClientServerCommandLineParser.java
@@ -14,13 +14,11 @@
 package megamek.common.commandline;
 
 import megamek.MMConstants;
-import megamek.MMOptions;
 import megamek.MegaMek;
 import megamek.client.ui.Messages;
 import megamek.common.Configuration;
 import megamek.common.annotations.Nullable;
-import megamek.common.preference.PreferenceManager;
-import megamek.server.Server;
+import megamek.server.Processor.MovementProcessor;
 import org.apache.logging.log4j.LogManager;
 
 import java.io.File;
@@ -196,7 +194,7 @@ public class ClientServerCommandLineParser extends AbstractCommandLineParser {
                         "port number must be a number. '%s' is not valid\n%s",
                         getTokenValue(), ex.getMessage()));
             }
-            port = Server.validatePort(newPort);
+            port = MovementProcessor.validatePort(newPort);
         } else {
             throw new ParseException("port number expected");
         }
@@ -212,7 +210,7 @@ public class ClientServerCommandLineParser extends AbstractCommandLineParser {
 
     private void parsePassword() throws ParseException {
         if (getTokenType() == TOK_LITERAL) {
-            password = Server.validatePassword(getTokenValue());
+            password = MovementProcessor.validatePassword(getTokenValue());
         } else {
             throw new ParseException("password expected");
         }
@@ -228,7 +226,7 @@ public class ClientServerCommandLineParser extends AbstractCommandLineParser {
 
     private void parsePlayerName() throws ParseException {
         if (getTokenType() == TOK_LITERAL) {
-            playerName = Server.validatePlayerName(getTokenValue());
+            playerName = MovementProcessor.validatePlayerName(getTokenValue());
         } else {
             throw new ParseException("playerName expected");
         }
@@ -236,7 +234,7 @@ public class ClientServerCommandLineParser extends AbstractCommandLineParser {
 
     private void parseServerAddress() throws ParseException {
         if (getTokenType() == TOK_LITERAL) {
-            serverAddress = Server.validateServerAddress(getTokenValue());
+            serverAddress = MovementProcessor.validateServerAddress(getTokenValue());
         } else {
             throw new ParseException("host name or url expected");
         }
